requirements.txt
Django==5.1
psycopg2-binary==2.9.9
djangorestframework==3.15.2
dj-database-url==2.2.0
gunicorn==22.0.0
django-filter==24.3
drf-spectacular==0.27.2
whitenoise==6.7.0
python-decouple==3.8
